requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
decorator==4.4.2
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
imageio==2.10.1
imageio-ffmpeg==0.4.5
moviepy==1.0.3
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
Pillow==8.4.0
praw==7.4.0
prawcore==2.3.0
proglog==0.1.9
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
tqdm==4.62.3
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.1
